[jnc_ct] fix: LlvmIrBuilder::createSwitch (intptr_t -> int64_t)

+++ src/jnc_ct/jnc_ct_Module/jnc_ct_LlvmIrBuilder.cpp
@@ -74,7 +74,7 @@ LlvmIrBuilder::createSwitch(
 		caseCount
 	);
 
-	sl::HashTableIterator<intptr_t, BasicBlock*> caseIt = firstCase;
+	sl::HashTableIterator<int64_t, BasicBlock*> caseIt = firstCase;
 	for (; caseIt; caseIt++) {
 		Value constValue(caseIt->getKey(), type);
 		BasicBlock* block = caseIt->m_value;
